backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.5.0
pymongo==3.12.1
pytz==2022.2.1
sqlparse==0.2.4
